# Conflicts:
#	Firebird.2.5/Tests/Linq/DateOnlyFunctionTests/Tests.Linq.DateOnlyFunctionTests.SubDateDay(Firebird.2.5).sql
#	Firebird.2.5/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateDay(Firebird.2.5).sql
#	Firebird.2.5/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateHour(Firebird.2.5).sql
#	Firebird.2.5/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMillisecond(Firebird.2.5).sql
#	Firebird.2.5/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMinute(Firebird.2.5).sql
#	Firebird.2.5/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(Firebird.2.5).sql
#	Firebird.3/Tests/Linq/DateOnlyFunctionTests/Tests.Linq.DateOnlyFunctionTests.SubDateDay(Firebird.3).sql
#	Firebird.3/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateDay(Firebird.3).sql
#	Firebird.3/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateHour(Firebird.3).sql
#	Firebird.3/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMillisecond(Firebird.3).sql
#	Firebird.3/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMinute(Firebird.3).sql
#	Firebird.3/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(Firebird.3).sql
#	Firebird.4/Tests/Linq/DateOnlyFunctionTests/Tests.Linq.DateOnlyFunctionTests.SubDateDay(Firebird.4).sql
#	Firebird.4/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateDay(Firebird.4).sql
#	Firebird.4/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateHour(Firebird.4).sql
#	Firebird.4/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMillisecond(Firebird.4).sql
#	Firebird.4/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMinute(Firebird.4).sql
#	Firebird.4/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(Firebird.4).sql
#	Firebird.5/Tests/Linq/DateOnlyFunctionTests/Tests.Linq.DateOnlyFunctionTests.SubDateDay(Firebird.5).sql
#	Firebird.5/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateDay(Firebird.5).sql
#	Firebird.5/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateHour(Firebird.5).sql
#	Firebird.5/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMillisecond(Firebird.5).sql
#	Firebird.5/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMinute(Firebird.5).sql
#	Firebird.5/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(Firebird.5).sql
#	Oracle.11.Managed/Tests/Linq/DateOnlyFunctionTests/Tests.Linq.DateOnlyFunctionTests.SubDateDay(Oracle.11.Managed).sql
#	Oracle.11.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateDay(Oracle.11.Managed).sql
#	Oracle.11.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateHour(Oracle.11.Managed).sql
#	Oracle.11.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMinute(Oracle.11.Managed).sql
#	Oracle.11.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(Oracle.11.Managed).sql
#	Oracle.12.Managed/Tests/Linq/DateOnlyFunctionTests/Tests.Linq.DateOnlyFunctionTests.SubDateDay(Oracle.12.Managed).sql
#	Oracle.12.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateDay(Oracle.12.Managed).sql
#	Oracle.12.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateHour(Oracle.12.Managed).sql
#	Oracle.12.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMinute(Oracle.12.Managed).sql
#	Oracle.12.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(Oracle.12.Managed).sql
#	Oracle.18.Managed/Tests/Linq/DateOnlyFunctionTests/Tests.Linq.DateOnlyFunctionTests.SubDateDay(Oracle.18.Managed).sql
#	Oracle.18.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateDay(Oracle.18.Managed).sql
#	Oracle.18.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateHour(Oracle.18.Managed).sql
#	Oracle.18.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMinute(Oracle.18.Managed).sql
#	Oracle.18.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(Oracle.18.Managed).sql
#	Oracle.19.Managed/Tests/Linq/DateOnlyFunctionTests/Tests.Linq.DateOnlyFunctionTests.SubDateDay(Oracle.19.Managed).sql
#	Oracle.19.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateDay(Oracle.19.Managed).sql
#	Oracle.19.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateHour(Oracle.19.Managed).sql
#	Oracle.19.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMinute(Oracle.19.Managed).sql
#	Oracle.19.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(Oracle.19.Managed).sql
#	Oracle.21.Managed/Tests/Linq/DateOnlyFunctionTests/Tests.Linq.DateOnlyFunctionTests.SubDateDay(Oracle.21.Managed).sql
#	Oracle.21.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateDay(Oracle.21.Managed).sql
#	Oracle.21.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateHour(Oracle.21.Managed).sql
#	Oracle.21.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMinute(Oracle.21.Managed).sql
#	Oracle.21.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(Oracle.21.Managed).sql
#	Oracle.23.Managed/Tests/Linq/DateOnlyFunctionTests/Tests.Linq.DateOnlyFunctionTests.SubDateDay(Oracle.23.Managed).sql
#	Oracle.23.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateDay(Oracle.23.Managed).sql
#	Oracle.23.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateHour(Oracle.23.Managed).sql
#	Oracle.23.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMinute(Oracle.23.Managed).sql
#	Oracle.23.Managed/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(Oracle.23.Managed).sql

+++ PostgreSQL.13/Tests/Linq/DateTimeOffsetTests/Tests.Linq.DateTimeOffsetTests.SubDateDay(PostgreSQL.13).sql
@@ -1,7 +1,7 @@
 ﻿-- PostgreSQL.13 PostgreSQL
 
 SELECT
-	EXTRACT(EPOCH FROM ((t."TransactionDate" + 96 * Interval '1 Hour')::timestamp - t."TransactionDate"::timestamp)) / 86400
+	(EXTRACT(EPOCH FROM ((t."TransactionDate" + 96 * Interval '1 Hour')::timestamp - t."TransactionDate"::timestamp)) / 86400)::Float
 FROM
 	"Transactions" t
